+++ src/main/java/com/ssafy/questory/controller/FollowController.java
@@ -3,7 +3,9 @@ package com.ssafy.questory.controller;
 import com.ssafy.questory.domain.Member;
 import com.ssafy.questory.dto.request.friend.FriendStatusRequestDto;
 import com.ssafy.questory.dto.request.member.MemberEmailRequestDto;
+import com.ssafy.questory.dto.request.member.MemberSearchRequestDto;
 import com.ssafy.questory.dto.response.friend.FollowResponseDto;
+import com.ssafy.questory.dto.response.member.MemberSearchResponseDto;
 import com.ssafy.questory.dto.response.member.auth.MemberInfoResponseDto;
 import com.ssafy.questory.service.FriendService;
 import io.swagger.v3.oas.annotations.Operation;
@@ -78,5 +80,13 @@ public class FollowController {
             @RequestParam(defaultValue = "0") int page,
             @RequestParam(defaultValue = "10") int size) {
         return ResponseEntity.ok(friendService.getFollowRequests(member, page, size));
+    }
+
+    @GetMapping("/search")
+    @Operation(summary = "유저 검색", description = "이메일을 통해 유저를 검색합니다.")
+    public ResponseEntity<Page<MemberSearchResponseDto>> search(
+            @AuthenticationPrincipal(expression = "member") Member member,
+            @ModelAttribute MemberSearchRequestDto memberSearchRequestDto) {
+        return ResponseEntity.ok(friendService.search(member, memberSearchRequestDto));
     }
 }
