+++ src/main/java/com/ssafy/questory/controller/MemberAuthController.java
@@ -58,12 +58,4 @@ public class MemberAuthController {
                 "message", "탈퇴하였습니다."
         ));
     }
-
-    @GetMapping("/search")
-    @Operation(summary = "유저 검색", description = "이메일을 통해 유저를 검색합니다.")
-    public ResponseEntity<Page<MemberSearchResponseDto>> search(
-            @AuthenticationPrincipal(expression = "member") Member member,
-            @ModelAttribute MemberSearchRequestDto memberSearchRequestDto) {
-        return ResponseEntity.ok(memberAuthService.search(member, memberSearchRequestDto));
-    }
 }
